file refactoring using traits
